[#5] Refactor: Modal Logic 분리
